html/browsers/history/the-history-interface/combination_history_006.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>combination_history_006(After calling of go method, check length and index.)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    var t = async_test("After calling of go method, check length and index");

    var last_length;
    var last_index;
    t.step(function () {
      window.history.pushState(1, document.title, '?x=1');
      window.history.pushState(2, document.title, '?x=2');

      last_length = window.history.length;
      last_index = window.history.index;

      window.history.go(-1);
    });

    window.addEventListener('popstate', t.step_func(function(e) {
      assert_equals(e.state, 1, "state");
      assert_equals(window.history.length, last_length, "last_length");
      assert_equals(window.history.index, last_index - 1, "last_index - 1");
      t.done();
    }), false);
    </script>
  </body>
</html>
